Add Reviews to database
Reviews (STI) will be of two types: review of mechanic or review of cyclist. Both types of reviews must 'belong to' both a cyclist and a mechanic (the author and the user that the review describes).

There are three fields that are author-dependent: if a cyclist is leaving a review for a mechanic, the review must indicate whether the service was performed as expected (boolean) and whether the price charged was as expected (boolean). If a mechanic is leaving a review for a cyclist, they must indicate whether or not the cyclist paid (boolean).

Reviews must also include a rating (integer, range tbd), and may optionally include comments.