Add RequestedServices to database
Requested Services acts as a join table between bookings and service, to allow both cyclists and mechanics to track the history of jobs done. Each booking has one or more services (eg, a wheel true and a brake adjustment), and each service can be performed several times by the same mechanic (eg, a wheel true for Jamie and a wheel true for Kelly - the service must be linked to each booking independently).