User email: spec, validation, & factory update
Add a basic spec that tests for the presence of an email when creating an instance of the User class. Add a validation to the User class definition to make the spec pass. Update the Users factory to use Faker to generate an email address when generating instances of the User class.